+++ apps/web/package.json
@@ -6,7 +6,7 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "preview": "vite preview",
-    "test": "vitest run src/appData/domain/index.test.ts src/appData/session/useWorkspaceSession.test.tsx src/appData/sync/syncLifecycleObservation.test.ts src/localDb/core/core.migrations.test.ts src/localDb/reviews/reviews.test.ts src/localDb/progress/progress.test.ts src/localDb/reviews/reviewSchedule.test.ts src/localDb/cards/cards.test.ts src/localDb/feedback/feedback.test.ts src/feedback/**/*.test.ts src/screens/review/**/*.test.ts src/screens/review/**/*.test.tsx src/screens/settings/**/*.test.tsx src/screens/progress/ProgressScreen.render.test.tsx src/appData/progress/source/progressSource.summarySeries.test.tsx src/appData/progress/source/progressSource.cache.test.tsx src/appData/progress/source/progressSource.reviewSchedule.test.tsx src/appData/progress/source/progressSource.lifecycle.test.tsx src/appData/progress/invalidation/progressInvalidation.test.tsx src/progress/progressDates.test.ts src/api/endpoints.test.ts src/api/transport.test.ts src/chat/streaming/liveStream.test.ts src/chat/attachments/*.test.ts src/chat/ChatPanel/tests/*.test.tsx src/chat/composer/*.test.ts src/accountDeletion.test.ts src/observability/instrument.test.ts",
+    "test": "vitest run src/appData/domain/index.test.ts src/appData/session/useWorkspaceSession.test.tsx src/appData/sync/observation/syncLifecycleObservation.test.ts src/localDb/core/core.migrations.test.ts src/localDb/reviews/reviews.test.ts src/localDb/progress/progress.test.ts src/localDb/reviews/reviewSchedule.test.ts src/localDb/cards/cards.test.ts src/localDb/feedback/feedback.test.ts src/feedback/**/*.test.ts src/screens/review/**/*.test.ts src/screens/review/**/*.test.tsx src/screens/settings/**/*.test.tsx src/screens/progress/ProgressScreen.render.test.tsx src/appData/progress/source/progressSource.summarySeries.test.tsx src/appData/progress/source/progressSource.cache.test.tsx src/appData/progress/source/progressSource.reviewSchedule.test.tsx src/appData/progress/source/progressSource.lifecycle.test.tsx src/appData/progress/invalidation/progressInvalidation.test.tsx src/progress/progressDates.test.ts src/api/endpoints.test.ts src/api/transport.test.ts src/chat/streaming/liveStream.test.ts src/chat/attachments/*.test.ts src/chat/ChatPanel/tests/*.test.tsx src/chat/composer/*.test.ts src/accountDeletion.test.ts src/observability/instrument.test.ts",
     "test:e2e": "FLASHCARDS_E2E_TARGET=prod playwright test",
     "test:e2e:local": "node ./scripts/check-local-e2e-stack.mjs && FLASHCARDS_E2E_TARGET=local playwright test",
     "test:e2e:prod": "FLASHCARDS_E2E_TARGET=prod playwright test"
