+++ apps/web/src/accountDeletion.test.ts
@@ -11,7 +11,7 @@ import { INSTALLATION_ID_STORAGE_KEY } from "./clientIdentity";
 import { LOCALE_PREFERENCE_STORAGE_KEY } from "./i18n/runtime";
 import { loadCloudSettings, putCloudSettings } from "./localDb/sync/cloudSettings";
 import { clearWebSyncCache } from "./localDb/cache";
-import { SYNC_RESTORE_HISTORY_STORAGE_KEY } from "./appData/sync/syncRestoreHistory";
+import { SYNC_RESTORE_HISTORY_STORAGE_KEY } from "./appData/sync/restore/syncRestoreHistory";
 import type { CloudSettings } from "./types";
 
 const observabilityMocks = vi.hoisted(() => ({
