+++ apps/web/src/appData/sync/local/syncCloudSettings.ts
@@ -1,0 +1,18 @@
+import { loadCloudSettings } from "../../../localDb/sync/cloudSettings";
+import type { CloudSettings } from "../../../types";
+
+export function requireCloudInstallationId(cloudSettings: CloudSettings | null): string {
+  if (cloudSettings === null) {
+    throw new Error("Cloud settings are not loaded");
+  }
+
+  if (cloudSettings.installationId.trim() === "") {
+    throw new Error("Cloud settings installationId is not loaded");
+  }
+
+  return cloudSettings.installationId;
+}
+
+export async function loadRequiredCloudInstallationId(): Promise<string> {
+  return requireCloudInstallationId(await loadCloudSettings());
+}
