+++ apps/web/src/appData/sync/remote/constants.ts
@@ -1,0 +1,2 @@
+export const syncPageSize = 500;
+export const slowHotBootstrapWarningThresholdMs = 2000;
